requirements.txt
# 藏鋒系統依賴清單

# 資料處理
pandas>=2.0.0
numpy>=1.24.0

# API 請求
requests>=2.31.0

# 資料庫
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Google Sheet
gspread>=5.12.0
google-auth>=2.23.0

# 排程
schedule>=1.2.0

# 環境變數
python-dotenv>=1.0.0

# 日誌與效能
loguru>=0.7.0

# 開發工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
